+++ ethik/cache_explainer.py
@@ -3,7 +3,6 @@ import functools
 import itertools
 import warnings
 
-import numpy as np
 import pandas as pd
 import plotly.graph_objs as go
 
@@ -41,7 +40,7 @@ class CacheExplainer(BaseExplainer):
         tol (float): The bottom threshold for the gradient of the optimization
             procedure. When reached, the procedure stops. Otherwise, a warning
             is raised about the fact that the optimization did not converge.
-            Default is `1e-3`.
+            Default is `1e-4`.
         n_jobs (int): The number of jobs to use for parallel computations. See
             `joblib.Parallel()`. Default is `-1`.
         memoize (bool): Indicates whether or not memoization should be used or not. If `True`, then
@@ -62,7 +61,7 @@ class CacheExplainer(BaseExplainer):
         sample_frac=0.8,
         conf_level=0.05,
         max_iterations=15,
-        tol=1e-3,
+        tol=1e-4,
         n_jobs=1,  # Parallelism is only worth it if the dataset is "large"
         memoize=False,
         verbose=True,